+++ .gitignore
@@ -20,6 +20,7 @@ xcuserdata/
 ## Other
 *.moved-aside
 *.xcuserstate
+*.xcworkspacedata
 
 ## Obj-C/Swift specific
 *.hmap
@@ -62,7 +63,3 @@ fastlane/report.xml
 fastlane/Preview.html
 fastlane/screenshots
 fastlane/test_output
-
-# Jazzy Documentation
-
-docs/
