+++ SKCollectionViewDataSource.podspec
@@ -1,8 +1,8 @@
 Pod::Spec.new do |spec|
   spec.name = 'SKCollectionViewDataSource'
-  spec.version = '1.0.0'
+  spec.version = '0.0.2'
   spec.license = 'MIT'
-  spec.summary = 'An easy to configure data source for UICollectionView.'
+  spec.summary = 'An easier to configure data source for UICollectionView.'
   spec.homepage = 'https://github.com/skladek/SKCollectionViewDataSource'
   spec.authors = { 'Sean Kladek' => 'skladek@gmail.com' }
   spec.source = { :git => 'https://github.com/skladek/SKCollectionViewDataSource.git', :tag => spec.version }
