+++ SampleProject/SampleProject/ViewControllers/HomeViewController.swift
@@ -25,6 +25,11 @@ class HomeViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        let reuseId = "HomeViewControllerReuseId"
+
+        let textCellNib = UINib(nibName: "TextCell", bundle: Bundle.main)
+        collectionView.register(textCellNib, forCellWithReuseIdentifier: reuseId)
+
         let array: [Items] = [
             .singleSection,
             .multipleSection,
@@ -32,8 +37,7 @@ class HomeViewController: UIViewController {
             .supplementaryViews
         ]
 
-        let textCellNib = UINib(nibName: "TextCell", bundle: Bundle.main)
-        let cellConfiguration = CellConfiguration<Items>(cell: textCellNib) { (cell, object) in
+        let cellConfiguration = CellConfiguration<Items>(reuseId: reuseId) { (cell, object) in
             guard let cell = cell as? TextCell else {
                 return
             }
