+++ SampleProject/SampleProject/ViewControllers/SupplementaryViewsViewController.swift
@@ -18,10 +18,7 @@ class SupplementaryViewsViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        let array = [["Alaska", "Alabama", "Arkansas", "American Samoa", "Arizona"], ["California", "Colorado", "Connecticut"], ["District of Columbia", "Delaware"], ["Florida"], ["Georgia", "Guam"], ["Hawaii"], ["Iowa", "Idaho", "Illinois", "Indiana"], ["Kansas", "Kentucky"], ["Louisiana"], ["Massachusetts", "Maryland", "Maine", "Michigan", "Minnesota", "Missouri", "Mississippi", "Montana"], ["North Carolina", "North Dakota", "Nebraska", "New Hampshire", "New Jersey", "New Mexico", "Nevada", "New York"], ["Ohio", "Oklahoma", "Oregon"], ["Pennsylvania", "Puerto Rico"], ["Rhode Island"], ["South Carolina", "South Dakota"], ["Tennessee", "Texas"], ["Utah"], ["Virginia", "Virgin Islands", "Vermont"], ["Washington", "Wisconsin", "West Virginia", "Wyoming"]]
-
-        let textCellNib = UINib(nibName: "TextCell", bundle: Bundle.main)
-        let cellConfiguration = CellConfiguration<String>(cell: textCellNib) { (cell, object) in
+        let cellConfiguration = CellConfiguration<String>(reuseId: "SupplementaryViewsViewControllerReuseId") { (cell, object) in
             guard let cell = cell as? TextCell else {
                 return
             }
@@ -29,8 +26,7 @@ class SupplementaryViewsViewController: UIViewController {
             cell.label.text = object
         }
 
-        let headerNib = UINib(nibName: "HeaderCell", bundle: Bundle.main)
-        let headerConfiguration = SupplementaryViewConfiguration<String>(view: headerNib, viewKind: UICollectionElementKindSectionHeader) { (view, section) in
+        let headerConfiguration = SupplementaryViewConfiguration<String>(reuseId: "HeaderViewReuseId", viewKind: UICollectionElementKindSectionHeader) { (view, section) in
             guard let view = view as? HeaderCell else {
                 return
             }
@@ -44,14 +40,24 @@ class SupplementaryViewsViewController: UIViewController {
             view.label.text = firstLetter
         }
 
-        let footerNib = UINib(nibName: "FooterCell", bundle: Bundle.main)
-        let footerConfiguration = SupplementaryViewConfiguration<String>(view: footerNib, viewKind: UICollectionElementKindSectionFooter) { (view, section) in
+        let footerConfiguration = SupplementaryViewConfiguration<String>(reuseId: "FooterViewReuseId", viewKind: UICollectionElementKindSectionFooter) { (view, section) in
             guard let view = view as? FooterCell else {
                 return
             }
 
             view.backgroundColor = .red
         }
+
+        let textCellNib = UINib(nibName: "TextCell", bundle: Bundle.main)
+        collectionView.register(textCellNib, forCellWithReuseIdentifier: cellConfiguration.reuseId)
+
+        let headerNib = UINib(nibName: "HeaderCell", bundle: Bundle.main)
+        collectionView.register(headerNib, forSupplementaryViewOfKind: headerConfiguration.viewKind, withReuseIdentifier: headerConfiguration.reuseId)
+
+        let footerNib = UINib(nibName: "FooterCell", bundle: Bundle.main)
+        collectionView.register(footerNib, forSupplementaryViewOfKind: footerConfiguration.viewKind, withReuseIdentifier: footerConfiguration.reuseId)
+
+        let array = [["Alaska", "Alabama", "Arkansas", "American Samoa", "Arizona"], ["California", "Colorado", "Connecticut"], ["District of Columbia", "Delaware"], ["Florida"], ["Georgia", "Guam"], ["Hawaii"], ["Iowa", "Idaho", "Illinois", "Indiana"], ["Kansas", "Kentucky"], ["Louisiana"], ["Massachusetts", "Maryland", "Maine", "Michigan", "Minnesota", "Missouri", "Mississippi", "Montana"], ["North Carolina", "North Dakota", "Nebraska", "New Hampshire", "New Jersey", "New Mexico", "Nevada", "New York"], ["Ohio", "Oklahoma", "Oregon"], ["Pennsylvania", "Puerto Rico"], ["Rhode Island"], ["South Carolina", "South Dakota"], ["Tennessee", "Texas"], ["Utah"], ["Virginia", "Virgin Islands", "Vermont"], ["Washington", "Wisconsin", "West Virginia", "Wyoming"]]
 
         dataSource = CollectionViewDataSource(objects: array, cellConfiguration: cellConfiguration, supplementaryViewConfigurations: [headerConfiguration, footerConfiguration])
