--- codecov.yml
@@ -1,3 +1,0 @@
-ignore:
-  - "Pods" 
-  - "Tests"
